engine/src/main/java/com/freedom/backend/engine/validator/ModelValidator.java
package com.freedom.backend.engine.validator;

import com.freedom.backend.engine.common.ErrorEnum;
import com.freedom.backend.engine.exception.DefinitionException;
import com.freedom.backend.engine.exception.ProcessException;
import com.freedom.backend.engine.model.FlowModel;
import com.freedom.backend.engine.param.CommonParam;
import com.freedom.backend.engine.util.FlowModelUtil;
import org.apache.commons.collections4.CollectionUtils;
import org.apache.commons.lang3.StringUtils;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Component;

import javax.annotation.Resource;

@Component
public class ModelValidator {

    private static final Logger LOGGER = LoggerFactory.getLogger(ModelValidator.class);

    @Resource
    private FlowModelValidator flowModelValidator;

    public void validate(String flowModelStr) throws DefinitionException, ProcessException {
        this.validate(flowModelStr, null);
    }

    public void validate(String flowModelStr, CommonParam commonParam) throws DefinitionException, ProcessException {
        if (StringUtils.isBlank(flowModelStr)) {
            LOGGER.warn("message={}", ErrorEnum.MODEL_EMPTY.getErrMsg());
            throw new DefinitionException(ErrorEnum.MODEL_EMPTY);
        }

        FlowModel flowModel = FlowModelUtil.parseModelFromString(flowModelStr);
        if (flowModel == null || CollectionUtils.isEmpty(flowModel.getFlowElementList())) {
            LOGGER.warn("message={}||flowModelStr={}", ErrorEnum.MODEL_EMPTY.getErrMsg(), flowModelStr);
            throw new DefinitionException(ErrorEnum.MODEL_EMPTY);
        }
        flowModelValidator.validate(flowModel, commonParam);
    }
}
